* master:
  GAME: Add `GameLoop::ReloadContext()`
  GAME: Lerp camera zoom
  GAME: Split `LayerID::FX`
  GAME: Add Camera zoom
  GFX: Add `GFX` to centralise coordinate spaces.
  REFACTOR: Let `GameManager` own WSM
  REFACTOR: Strip constant `viewSize` assumption.
  REFACTOR: Use `enum class : T` instead of `enum`.
  CMAKE: `ThirdParty` minor refactor

+++ Source/Engine/Engine/Physics/Collider.cpp
@@ -52,7 +52,7 @@ Collider::~Collider() = default;
 #if ENABLED(DEBUG_LOGGING)
 void Collider::OnHit(const Collider& other)
 {
-	static int DEBUG_skip = 0;
+	static u8 DEBUG_skip = 0;
 	if (++DEBUG_skip > 20)
 	{
 		LOG_D("[%s] colliding with [%s] (throttled log)", m_name.c_str(), other.m_name.c_str());
